Edit 103 file

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -11,7 +11,7 @@
 
 int main(void)
 {
-	int i;
+	long long int i;
 	long int next;
 	long int prev;
 	long int sum;
